pyproject.toml
[tool.poetry]
name = "django-allauth-webauthn"
version = "0.1.0"
description = ""
authors = ["H. Gregor Molter <gregor.molter@secretlab.de>"]
readme = "README.rst"
license = "MIT"
homepage = "https://github.com/cyberdeck/django-allauth-webauthn"
repository = "https://github.com/cyberdeck/django-allauth-webauthn"
documentation = "https://django-allauth-webauthn.readthedocs.io/en/latest/"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
exclude = [
    "demo/"
]

[tool.poetry.urls]
'Bug Tracker' = 'https://github.com/cyberdeck/django-allauth-webauthn/issues'

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
django-allauth = "^0.45.0"
webauthn = "^0.4.7"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.3"
pytest = "^5.2"
Werkzeug = "^2.0.1"
coverage = {extras = ["toml"], version = "^5.5"}
flake8 = "^3.9.2"
flake8-bugbear = "^21.9.1"
mypy = "^0.910"
pre-commit = "^2.15.0"
Sphinx = "^4.2.0"
sphinx-autobuild = "^2021.3.14"
sphinx-rtd-theme = "^1.0.0"
toml = "^0.10.2"

[tool.coverage.run]
branch = true
omit = [".*", "*/site-packages/*", "manage.py", "tests/*"]

[tool.coverage.report]
fail_under = 100

[tool.coverage.html]
directory = "coverage_report"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
